changed to use int64_t rather than int. refs #349
git-svn-id: https://svn.nexusformat.org/code/branches/4.3@1898 ff5d1e40-2be0-497f-93bd-dc18237bd3c7

+++ src/napi5.c
@@ -1906,7 +1906,7 @@ static int countObjectsInGroup(hid_t loc_id)
      *rank = (int)iRank;
      for (i = 0; i < iRank; i++)
      {
-	  dimension[i] = (int)myDim[i];
+	  dimension[i] = (int64_t)myDim[i];
      }
      return NX_OK;
    }
